refactor(image): decoupling browser as a new component image-browser
