+++ packages/image-browser/image-browser-icon.tsx
