+++ packages/image-browser/index.less
@@ -1,0 +1,105 @@
+.fect-image__browser {
+  background-color: transparent;
+  box-shadow: var(--fay-shadowLarge);
+  width: max-content;
+  max-width: 100%;
+  margin: 0 auto;
+  border-radius: var(--fay-radius);
+  overflow: hidden;
+  .fect-image {
+    border-top-left-radius: 0;
+    border-top-right-radius: 0;
+  }
+  header {
+    height: 40px;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    border-bottom: 1px solid var(--accents-2);
+
+    .traffic {
+      width: auto;
+      position: absolute;
+      left: var(--fay-gap-half);
+      top: 50%;
+      transform: translateY(-50%);
+      bottom: 0;
+      height: 100%;
+      display: flex;
+      align-items: center;
+      user-select: none;
+      span {
+        border-radius: 50%;
+        width: 12px;
+        height: 12px;
+        display: inline-block;
+        margin-right: 8px;
+      }
+      .close {
+        background-color: #ff5f56;
+      }
+
+      .mini {
+        background-color: #ffbd2e;
+      }
+
+      .full {
+        background-color: #27c93f;
+      }
+    }
+
+    .title {
+      font-size: 12px;
+    }
+  }
+}
+
+.fect-image__browser {
+  .address-input {
+    height: 28px;
+    min-width: 40%;
+    max-width: 60%;
+    color: inherit;
+    border-radius: 3px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0 10px;
+    overflow: hidden;
+    position: relative;
+    font-size: 12px;
+    color: inherit;
+    a {
+      max-width: 90%;
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+      display: inline-block;
+      color: inherit;
+    }
+    .https {
+      width: 12px;
+      height: 12px;
+      margin-right: 5px;
+      user-select: none;
+      margin-top: -1px;
+      color: inherit;
+    }
+  }
+}
+
+.fect-image {
+  border-top-left-radius: 0;
+  border-top-right-radius: 0;
+  &__browser {
+    background-color: transparent;
+    box-shadow: var(--fay-shadowLarge);
+    width: max-content;
+    max-width: 100%;
+    margin: 0 auto;
+    border-radius: var(--fay-radius);
+    overflow: hidden;
+  }
+}
