+++ packages/image-browser/type.ts
@@ -1,0 +1,7 @@
+export type BrowserColors = {
+  color: string
+  barBgColor: string
+  inputBgColor: string
+  borderColor: string
+  titleColor: string
+}
